DAO Classes Finished

+++ src/org/angello/com/model/entities/Cliente.java
@@ -2,14 +2,14 @@ package org.angello.com.model.entities;
 
 public class Cliente {
     // Attributes
-    private int CUI;
+    private Long CUI;
     private String nombre;
     private String apellido;
     private int telefono;
     private String email;
 
     // Constructor
-    public Cliente(int CUI, String nombre, String apellido, int telefono, String email) {
+    public Cliente(Long CUI, String nombre, String apellido, int telefono, String email) {
         this.CUI = CUI;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -18,7 +18,7 @@ public class Cliente {
     }
 
     // Getters
-    public int getCUI() {
+    public Long getCUI() {
         return CUI;
     }
 
@@ -39,7 +39,7 @@ public class Cliente {
     }
 
     // Setters
-    public void setCUI(int CUI) {
+    public void setCUI(Long CUI) {
         this.CUI = CUI;
     }
